centralize DRA version

+++ apps/basic.sh
@@ -1,22 +1,19 @@
-#!/bin/bash
+!/bin/bash
 
-echo ">>>> Update packages"
+echo ">>>> Basic packeges"
 sudo apt -q -y update
 sudo add-apt-repository universe
 sudo apt -q -y update
-
-echo ">>>> Install basic packeges"
 sudo apt -q -y install \
     build-essential autoconf linux-kernel-headers \
     curl wget tree whois unzip gpg htop \
     jq httpie bat fd-find gnome-tweaks dconf-cli \
-    fonts-firacode apt-utils ca-certificates gnupg \
-    neofetch
+    fonts-firacode apt-utils ca-certificates gnupg
 
-
-echo ">>>> Install basic packeges"
-curl -fsSL -o dra.deb \
-    https://github.com/devmatteini/dra/releases/download/0.3.8/dra_0.3.8_amd64.deb
-sudo dpkg -i ./dra.deb
-rm ./dra.deb
-dra completion bash >"$HOME"/.local/share/bash-completion/completions/dra
+if ! dra --version &>/dev/null; then
+  DRA_VERSION=0.4.2
+  curl -fsSL -o dra.deb "https://github.com/devmatteini/dra/releases/download/$DRA_VERSION/dra_${DRA_VERSION}_amd64.deb"
+  sudo dpkg -i ./dra.deb
+  rm ./dra.deb
+  dra completion bash >"$HOME"/.local/share/bash-completion/completions/dra
+fi
